Guerry-package(Guerry)	R Documentation

Guerry: maps data and methods related to A.-M. Guerry's 'Moral Statistics of France' (1833)

Description

Andre-Michel Guerry (1833) was the first to systematically collect and analyze social data on such things as crime, literacy and suicide with the view to determining social laws and the relations among these variables.

This package comprises maps of France in 1830, data from A.-M. Guerry and others, and statistical and graphic methods related to Guerry's "Moral Statistics of France".

The goal of providing these as an R package is to facilitate the exploration and development of statistical and graphic methods for multivariate data in a geo-spatial context.

Details

Package: 	 Guerry

Type: 	 Package

Version: 	 1.0

Date: 	 2009-10-20

License: 	 GPL

LazyLoad: 	 yes

Data from Guerry and others is contained in the data frame Guerry. Because Corsica is often considered an outlier both spatially and statistically, the map of France circa 1830, together with the Guerry data is provided as SpatialPolygonsDataFrames in two forms: gfrance for all 86 departments, and and gfrance85, for the 85 departments excluding Corsica.

Author(s)

Michael Friendly and Stephane Dray

Maintainer: Michael Friendly <friendly AT yorku.ca>

References

Dray, S. and Jombart, T. (2009). A Revisit Of Guerry's Data: Introducing Spatial Constraints In Multivariate Analysis. Unpublished manuscript.

Brunsdon, C. and Dykes, J. (2007). Geographically weighted visualization: interactive graphics for scale-varying exploratory analysis. Geographical Information Science Research Conference (GISRUK 2007). NUI Maynooth, Ireland, April, 2007. http://ncg.nuim.ie/gisruk/materials/proceedings/PDF/3C1.pdf.

Friendly, M. (2007). A.-M. Guerry's Moral Statistics of France: Challenges for Multivariable Spatial Analysis. Statistical Science, 22, 368-399.

Friendly, M. (2007). Supplementary materials for André-Michel Guerry's Moral Statistics of France: Challenges for Multivariate Spatial Analysis, http://www.math.yorku.ca/SCS/Gallery/guerry/.

Guerry, A.-M. (1833). Essai sur la statistique morale de la France Paris: Crochard. English translation: Hugh P. Whitt and Victor W. Reinking, Lewiston, N.Y.: Edwin Mellen Press, 2002.

Examples


#~~ simple examples of the most important functions ~~

[Package Guerry version 1.0 Index]