gfrance(Guerry)	R Documentation

Map of France in 1830 with the Guerry data

Description

gfrance is a SpatialPolygonsDataFrame object created with the sp package, containing the polygon boundaries of the map of France as it was in 1830, together with the Guerry data frame.

Usage

data(gfrance)

Format

The format is: Formal class 'SpatialPolygonsDataFrame' [package "sp"] with 5 slots: gfrance@data, gfrance@polygons, gfrance@plotOrder, gfrance@bbox, gfrance@proj4string. See: SpatialPolygonsDataFrame for descriptions of some components.

The analysis variables are described in Guerry.

Details

In the present version, the PROJ4 projection is not specified.

Source

Friendly, M. (2007). Supplementary materials for André-Michel Guerry's Moral Statistics of France: Challenges for Multivariate Spatial Analysis, http://www.math.yorku.ca/SCS/Gallery/guerry/.

References

Friendly, M. (2007). A.-M. Guerry's Moral Statistics of France: Challenges for Multivariable Spatial Analysis. Statistical Science, 22, 368-399.

Examples


data(gfrance)
names(gfrance)  ## list @data variables
plot(gfrance)   ## just show the map outline

# Show basic choropleth plots of some of the variables
spplot(gfrance, "Crime_pers")
spplot(gfrance, "Crime_prop")

# Note that spplot assumes all variables are on the same scale for comparative plots
# These plots do that, but it would be more appropriate to scale each separately, as in the above
spplot(gfrance, c("Crime_pers", "Crime_prop"))

[Package Guerry version 1.0 Index]